refactor: boolean decoder
